Error amendments
There was an issue where I had tested the code using an already defined
global version of a certain variable and function that allowed the code to run.
However, when running in a clean instance the code does not work (as expected).
This commit has amended that issue, specifically for `strata_age_binned_weight_proportions`.