Added `grouped_merge` method/function
An additional utility function `group_merge` has been
added to reduce the amount of repetition in cases where
multiple dataframes are being merged in the same step/pipeline/chain.
This doesn't change the previous output/result of the code, but it is
expected to be used for later calculations/steps that will enable
more consistent formatting and ensuring that the grouped merges are
being performed in the same way every time. This is particularly
important so the 'how' and 'on' arguments are appropriately applied
and are less vulnerable to errant typos.